preloadするリソースのテンプレート用変数を生成するコードを、ローカル・プラグインに分離
